agent: Add CreateContainer support for pre-pulled bundle
Add a check in setup_bundle to see if the bundle already exists and if it does then skip the setup.

The reason that k8s-kill-all-process-in-container.bats failed is that
deletion of the directory `/root/kata-containers/cid/rootfs` failed during removing container
because it was mounted twice (one in image-rs and one in set_bundle ) and only unmounted once in removing container.

Fix tests:
- k8s-credentials-secrets.bats
- k8s-file-volume.bats
- k8s-nested-configmap-secret.bats
- k8s-projected-volume.bats
- k8s-volume.bats
- k8s-shared-volume.bats
- k8s-sysctls.bats
- k8s-inotify.bats
- k8s-liveness-probes.bats
- k8s-kill-all-process-in-container.bats

Fixes: #9664
Fixes: #9665
Fixes: #9666
Fixes: #9667
Fixes: #9668

Signed-off-by: ChengyuZhu6 <chengyu.zhu@intel.com>
Co-authored-by: Wang, Arron <arron.wang@intel.com>
Co-authored-by: Dave Hay <david_hay@uk.ibm.com>
Co-authored-by: stevenhorsman <steven@uk.ibm.com>