runtime: detect the sealed secret volume
To avoid impacting other non-sealed secrets, user need ensure the volume name is prefixed with "sealed-secret" when using a sealed secret.
So that kata could detect sealed secret volumes by checking if the volume name is prefixed with "sealed-secret".

Due to the loss of mount name in OCI mounts, we need alternative methods to identify sealed secret mounts.
To address this, we currently add a "/sealed" prefix to the destination, marking the mount as a sealed-secret mount.
This allows the agent to locate and handle these mounts appropriately.

Signed-off-by: ChengyuZhu6 <chengyu.zhu@intel.com>
Signed-off-by: Linda Yu <linda.yu@intel.com>